Optimize `ctrl + d/u` actions
This commit optimizes (FINALLY!) the `ctrl` + `d/u` actions.

It does it by not printing 14 lines line-by-line, but it jumps quickly to the desired line, and if it isn't on screen, then it will quickly jump to the last visible edge, and then print line-by-line. This speeds up jumping a lot when details are on, but still relies on the speed of printing a line.

Printing a line could be made faster only by doing two things:
1. Heavily refactoring the whole `read_dir` and associated functions, to implement some sort of lazy loading for directories.
2. Make details more faster by using other tools for checking date and time. Less refactoring, but still some.